return back to double

+++ CMP1903M A01 2223/Interface1.cs
@@ -8,7 +8,7 @@ namespace CMP1903M_A01_2223
 {
     interface Ioperands
     {
-        float performance();
+        double performance();
         string ToString();
     }
 }
